feat(utility): Moved MultiProcessingCudaEnv from tests to modalities.
