+++ tests/training/test_activation_checkpointing.py
@@ -13,7 +13,7 @@ from modalities.__main__ import Main
 from modalities.config.config import ProcessGroupBackendType
 from modalities.config.pydantic_if_types import PydanticPytorchModuleType
 from modalities.models.gpt2.gpt2_model import GPT2Block
-from tests.end2end_tests.custom_components import MultiProcessingCudaEnv
+from modalities.running_env.cuda_env import MultiProcessingCudaEnv
 
 working_dir = Path(os.path.dirname(__file__))
